Can now write a file in fuse and read in JS and vice versa. Changed the fileproperties to use a long for filesize (emulated as two ints in JS). BOOYAAH!

+++ src/peergos/user/fs/FileProperties.java
@@ -26,7 +26,7 @@ public class FileProperties {
     public byte[] serialize() {
         DataSink dout = new DataSink();
         dout.writeString(name);
-        dout.writeDouble(size);
+        dout.writeLong(size);
         dout.writeDouble(modified.toEpochSecond(ZoneOffset.UTC));
         dout.writeBoolean(isHidden);
         if (!thumbnail.isPresent())
@@ -40,7 +40,7 @@ public class FileProperties {
     public static FileProperties deserialize(byte[] raw) throws IOException {
         DataSource din = new DataSource(raw);
         String name = din.readString();
-        long size = (long) din.readDouble();
+        long size = din.readLong();
         double modified = din.readDouble();
         boolean isHidden = din.readBoolean();
         int length = din.readInt();
